Optimizing code by adding separate transaction form component
